Undid a bad thing
First of all, I put that line that generates `__init__.py` files under the wrong section so I might have to fix some stuff with the golang SDK. Secondly, I decided to move the Python SDK from traditional packages to namespace packages. My first thought was that traditional packages would be best because they support both Python 2 and Python 3 but, given the prevalence of Python 3, the fact that namespace package support has been added to Python 2 retrospectively, and the fact that protobuf produces a directory structure that is incompatible with traditional packages, I made the decision to revert this change altogether.